+++ src/poa/contract/types.go
@@ -1,0 +1,7 @@
+package contract
+
+type solidityFields struct {
+	Constants string
+	AddTo     string
+	Checks    string
+}